requirements.txt
PyJWT==1.7.0
Django==2.2.4
